+++ tfts/models/informer.py
@@ -20,6 +20,8 @@ from tfts.layers.attention_layer import FullAttention, ProbAttention
 from tfts.layers.embed_layer import DataEmbedding, TokenEmbedding
 from tfts.layers.mask_layer import CausalMask
 
+from .base import BaseConfig, BaseModel
+
 params: Dict[str, Any] = {
     "n_encoder_layers": 1,
     "n_decoder_layers": 1,
