+++ tfts/models/seq2seq.py
@@ -13,6 +13,8 @@ from tensorflow.keras.layers import GRU, LSTM, RNN, Dense, Dropout, GRUCell, LST
 
 from tfts.layers.attention_layer import FullAttention
 
+from .base import BaseConfig, BaseModel
+
 params: Dict[str, Any] = {
     "rnn_type": "gru",
     "bi_direction": False,
@@ -27,11 +29,6 @@ params: Dict[str, Any] = {
     "skip_connect_circle": False,
     "skip_connect_mean": False,
 }
-
-
-@dataclass
-class ModelArguments:
-    pass
 
 
 class Seq2seq(object):
