+++ tfts/models/unet.py
@@ -10,6 +10,8 @@ from tensorflow.keras.layers import Activation, Add, AveragePooling1D, Concatena
 
 from tfts.layers.unet_layer import conv_br, re_block
 
+from .base import BaseConfig, BaseModel
+
 params: Dict[str, Any] = {
     "skip_connect_circle": False,
     "skip_connect_mean": False,
